// src/Clients/FaasNet.Website/src/stores/vpn/models/application.model.ts
import { ApplicationLinkResult } from "./applicationlink.model";

export class ApplicationResult {
  id: string = "";
  clientId: string | null = null;
  title: string = "";
  description: string = "";
  version: number = 0;
  posX: number = 0;
  posY: number = 0;
  links: ApplicationLinkResult[] = [];
}